7. Damage and Sync HP

+++ Assets/Scripts/PlayerWeapon.cs
@@ -5,6 +5,6 @@ using UnityEngine;
 [System.Serializable]
 public class PlayerWeapon {
     public string Name = "Glock";
-    public float Damage = 10f;
+    public int Damage = 10;
     public float Range = 100f;
 }
